requirements.txt
Pillow==7.0.0
pyusb==1.0.2
